Refactor, Add Move Mock, Rename action custom week details button
