+++ app/src/main/java/de/lukaspanni/opensourcestats/ui/details/DetailsFragment.java
@@ -11,6 +11,8 @@ import androidx.navigation.Navigation;
 
 import com.lukaspanni.opensourcestats.R;
 
+import java.util.Date;
+
 import de.lukaspanni.opensourcestats.data.TimeSpanFactory;
 
 
@@ -21,27 +23,28 @@ public class DetailsFragment extends Fragment {
 
         View view = inflater.inflate(R.layout.fragment_details, container, false);
 
+        
         view.findViewById(R.id.to_current_week_details).setOnClickListener(v -> {
             Bundle b = new Bundle();
             b.putParcelable("timeSpan", TimeSpanFactory.getCurrentWeek());
-            Navigation.findNavController(view).navigate(R.id.action_details_to_week_details, b);
+            Navigation.findNavController(view).navigate(R.id.action_details_to_time_span_details, b);
         });
         view.findViewById(R.id.to_last_week_details).setOnClickListener(v -> {
             Bundle b = new Bundle();
             b.putParcelable("timeSpan", TimeSpanFactory.getLastWeek());
-            Navigation.findNavController(view).navigate(R.id.action_details_to_week_details, b);
+            Navigation.findNavController(view).navigate(R.id.action_details_to_time_span_details, b);
         });
 
         view.findViewById(R.id.to_current_month_details).setOnClickListener(v -> {
             Bundle b = new Bundle();
             b.putParcelable("timeSpan", TimeSpanFactory.getCurrentMonth());
-            Navigation.findNavController(view).navigate(R.id.action_details_to_week_details, b);
+            Navigation.findNavController(view).navigate(R.id.action_details_to_time_span_details, b);
         });
 
         view.findViewById(R.id.to_last_month_details).setOnClickListener(v -> {
             Bundle b = new Bundle();
             b.putParcelable("timeSpan", TimeSpanFactory.getLastMonth());
-            Navigation.findNavController(view).navigate(R.id.action_details_to_week_details, b);
+            Navigation.findNavController(view).navigate(R.id.action_details_to_time_span_details, b);
         });
