+++ app/src/main/res/layout/fragment_details.xml
@@ -12,10 +12,26 @@
         android:layout_height="wrap_content">
 
         <Button
+            android:id="@+id/to_custom_week_details"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="30dp"
+            android:layout_marginRight="30dp"
+            android:layout_marginTop="40dp"
+            android:layout_weight="1"
+            android:background="@color/colorPrimary"
+            android:padding="20dp"
+            android:text="@string/to_week_details_text"
+            android:textAlignment="center"
+            android:textColor="@color/textWhite"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <Button
             android:id="@+id/to_current_week_details"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginLeft="30dp"
+            android:layout_marginTop="40dp"
             android:layout_marginRight="30dp"
             android:layout_weight="1"
             android:background="@color/colorPrimary"
@@ -23,7 +39,7 @@
             android:text="@string/to_current_week_details_text"
             android:textAlignment="center"
             android:textColor="@color/textWhite"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toBottomOf="@id/to_custom_week_details" />
 
         <Button
             android:id="@+id/to_last_week_details"
@@ -45,7 +61,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginLeft="30dp"
-            android:layout_marginTop="30dp"
+            android:layout_marginTop="40dp"
             android:layout_marginRight="30dp"
             android:layout_weight="1"
             android:background="@color/colorPrimary"
