+++ build.gradle
@@ -6,7 +6,7 @@ buildscript {
         jcenter()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:4.0.2"
+        classpath "com.android.tools.build:gradle:4.0.1"
         classpath "com.apollographql.apollo:apollo-gradle-plugin:2.4.0"
         classpath 'de.mobilej.unmock:UnMockPlugin:0.7.6'
     }
